Re-implement removal of genesis persistence in state db (#1295)
* Self-contained solution

* Retrieve new UT from #1017

* Keep changelog from #1017

* Updated changelog

* Appease linter

* Appease linter 2

---------
