+++ .changelog/unreleased/improvements/1017-remove-genesis-persistence-in-state-db.md
@@ -1,0 +1,3 @@
+- `[node]` Remove genesis persistence in state db, replaced by a hash
+  ([cometbft/cometbft\#1017](https://github.com/cometbft/cometbft/pull/1017),
+  [cometbft/cometbft\#1295](https://github.com/cometbft/cometbft/pull/1295))
